Only run push for merge jobs for docs
This sets us up to be able to run a verify-docs job

+++ jjb/nsh_sfc/include-raw-nsh_sfc-docs.sh
@@ -10,38 +10,40 @@ set -xe -o pipefail
 [ "$VERSION" ] || VERSION=$(./nsh-plugin/packaging/scripts/version)
 
 make doxygen
-mkdir -p $(dirname ${RESOURCES_DIR})
-mv -f ${DOC_DIR} ${RESOURCES_DIR}
-cd ${SITE_DIR}
-find . -type f '(' -name '*.md5' -o -name '*.dot' -o -name '*.map' ')' -delete
-cat > pom.xml << EOF
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>io.fd.${PROJECT_NAME}</groupId>
-  <artifactId>docs</artifactId>
-  <version>1.0.0</version>
-  <packaging>pom</packaging>
+if [[ ${JOB_NAME} == *merge* ]]; then
+  mkdir -p $(dirname ${RESOURCES_DIR})
+  mv -f ${DOC_DIR} ${RESOURCES_DIR}
+  cd ${SITE_DIR}
+  find . -type f '(' -name '*.md5' -o -name '*.dot' -o -name '*.map' ')' -delete
+  cat > pom.xml << EOF
+  <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>io.fd.${PROJECT_NAME}</groupId>
+    <artifactId>docs</artifactId>
+    <version>1.0.0</version>
+    <packaging>pom</packaging>
 
-  <properties>
-    <generateReports>false</generateReports>
-  </properties>
+    <properties>
+      <generateReports>false</generateReports>
+    </properties>
 
-  <build>
-    <extensions>
-      <extension>
-        <groupId>org.apache.maven.wagon</groupId>
-         <artifactId>wagon-webdav-jackrabbit</artifactId>
-         <version>2.9</version>
-      </extension>
-    </extensions>
-  </build>
-  <distributionManagement>
-    <site>
-      <id>fdio-site</id>
-      <url>dav:${DOCS_REPO_URL}/${PROJECT_PATH}/${VERSION}</url>
-    </site>
-  </distributionManagement>
-</project>
-EOF
-${MVN} site:site site:deploy -gs "${GLOBAL_SETTINGS_FILE}" -s "${SETTINGS_FILE}" -T 4C
-cd -
+    <build>
+      <extensions>
+        <extension>
+          <groupId>org.apache.maven.wagon</groupId>
+           <artifactId>wagon-webdav-jackrabbit</artifactId>
+           <version>2.9</version>
+        </extension>
+      </extensions>
+    </build>
+    <distributionManagement>
+      <site>
+        <id>fdio-site</id>
+        <url>dav:${DOCS_REPO_URL}/${PROJECT_PATH}/${VERSION}</url>
+      </site>
+    </distributionManagement>
+  </project>
+  EOF
+  ${MVN} site:site site:deploy -gs "${GLOBAL_SETTINGS_FILE}" -s "${SETTINGS_FILE}" -T 4C
+  cd -
+fi
